+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This file is used to list changes made in each version of the gem.
   - Exec Docker `CMD` in entrypoints.
   - Configure Puma and Redis.
   - Remove docker-compose.prod.
+- Version Docker generator.
 
 # 1.4.0.1